FujiSpot_backend/requirements.txt
flask
ultralytics
Pillow
scikit-learn
numpy